+++ src/torchmetrics/collections.py
@@ -14,7 +14,7 @@
 # this is just a bypass for this module name collision with build-in one
 from collections import OrderedDict
 from copy import deepcopy
-from typing import Any, Dict, Hashable, Iterable, List, Optional, Sequence, Tuple, Union
+from typing import Any, Dict, Hashable, Iterable, Iterator, List, Optional, Sequence, Tuple, Union
 
 import torch
 from torch import Tensor
@@ -150,6 +150,7 @@ class MetricCollection(ModuleDict):
          'valmetrics/MulticlassPrecision_micro': tensor(0.1250)}
     """
 
+    _modules: Dict[str, Metric]  # type: ignore[assignment]
     _groups: Dict[int, List[str]]
 
     def __init__(
@@ -187,7 +188,7 @@ class MetricCollection(ModuleDict):
         """
         # Use compute groups if already initialized and checked
         if self._groups_checked:
-            for _, cg in self._groups.items():
+            for cg in self._groups.values():
                 # only update the first member
                 m0 = getattr(self, cg[0])
                 m0.update(*args, **m0._filter_kwargs(**kwargs))
@@ -196,7 +197,7 @@ class MetricCollection(ModuleDict):
                 self._compute_groups_create_state_ref()
                 self._state_is_copy = False
         else:  # the first update always do per metric to form compute groups
-            for _, m in self.items(keep_base=True, copy_state=False):
+            for m in self.values(copy_state=False):
                 m_kwargs = m._filter_kwargs(**kwargs)
                 m.update(*args, **m_kwargs)
 
@@ -274,7 +275,7 @@ class MetricCollection(ModuleDict):
                 of just passed by reference
         """
         if not self._state_is_copy:
-            for _, cg in self._groups.items():
+            for cg in self._groups.values():
                 m0 = getattr(self, cg[0])
                 for i in range(1, len(cg)):
                     mi = getattr(self, cg[i])
@@ -327,7 +328,7 @@ class MetricCollection(ModuleDict):
 
     def reset(self) -> None:
         """Call reset for each metric sequentially."""
-        for _, m in self.items(keep_base=True, copy_state=False):
+        for m in self.values(copy_state=False):
             m.reset()
         if self._enable_compute_groups and self._groups_checked:
             # reset state reference
@@ -350,7 +351,7 @@ class MetricCollection(ModuleDict):
 
     def persistent(self, mode: bool = True) -> None:
         """Change if metric states should be saved to its state_dict after initialization."""
-        for _, m in self.items(keep_base=True, copy_state=False):
+        for m in self.values(copy_state=False):
             m.persistent(mode)
 
     def add_metrics(
@@ -459,6 +460,10 @@ class MetricCollection(ModuleDict):
             od[self._set_name(k)] = v
         return od
 
+    def __iter__(self) -> Iterator[Hashable]:
+        """Return an iterator over the keys of the MetricDict."""
+        return iter(self.keys())
+
     # TODO: redefine this as native python dict
     def keys(self, keep_base: bool = False) -> Iterable[Hashable]:
         r"""Return an iterable of the ModuleDict key.
@@ -470,7 +475,7 @@ class MetricCollection(ModuleDict):
             return self._modules.keys()
         return self._to_renamed_ordered_dict().keys()
 
-    def items(self, keep_base: bool = False, copy_state: bool = True) -> Iterable[Tuple[str, Module]]:
+    def items(self, keep_base: bool = False, copy_state: bool = True) -> Iterable[Tuple[str, Metric]]:
         r"""Return an iterable of the ModuleDict key/value pairs.
 
         Args:
@@ -483,7 +488,7 @@ class MetricCollection(ModuleDict):
             return self._modules.items()
         return self._to_renamed_ordered_dict().items()
 
-    def values(self, copy_state: bool = True) -> Iterable[Module]:
+    def values(self, copy_state: bool = True) -> Iterable[Metric]:
         """Return an iterable of the ModuleDict values.
 
         Args:
@@ -493,7 +498,7 @@ class MetricCollection(ModuleDict):
         self._compute_groups_create_state_ref(copy_state)
         return self._modules.values()
 
-    def __getitem__(self, key: str, copy_state: bool = True) -> Module:
+    def __getitem__(self, key: str, copy_state: bool = True) -> Metric:
         """Retrieve a single metric from the collection.
 
         Args:
@@ -525,7 +530,7 @@ class MetricCollection(ModuleDict):
         Arguments:
             dst_type (type or string): the desired type.
         """
-        for _, m in self.items(keep_base=True, copy_state=False):
+        for m in self.values(copy_state=False):
             m.set_dtype(dst_type)
         return self